requirements.txt
attrs==18.1.0
beautifulsoup4==4.6.0
bs4==0.0.1
cached-property==1.4.2
certifi==2018.4.16
chardet==3.0.4
click==6.7
configparser==3.5.0
Django==2.0.1
django-bootstrap4==0.0.2
django-crispy-forms==1.7.0
django-jet==1.0.7
django-suit==2.0a1
django-tables2==1.21.2
djangorestframework==3.8.2
enum34==1.1.6
feedparser==5.2.1
future==0.16.0
google-api-python-client==1.4.1
httmock==1.2.3
httplib2==0.11.3
idna==2.6
lxml==4.2.1
nose==1.3.7
oauth2client==4.1.2
oauthlib==2.0.7
ordereddict==1.1
parsedatetime==2.4
pathlib==1.0.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
python-dateutil==2.7.3
pytz==2017.3
requests==2.7.0
requests-oauthlib==0.8.0
rsa==3.4.2
schedule==0.5.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
WooCommerce==1.2.1
wordpress-api==1.2.6
zope.interface==4.5.0
